Added default stylings to the application and built a basic widget tree
Till now the application was ugly, there was nothing. We replace the hello world text widget with Scaffold widget, and we use it to style our application. Now, we have a widget tree, we have our overall MaterialApp widget, then we have Scaffold widget to get some basic page stylings, then we have AppBar widget in which we are only configuring a title with a Text widget, and on the same level we have a body which is also defined with a Text widget.